tmp/gemini/streamlit/pyproject.toml
[tool.poetry]
name = "prompt-googleai-streamlit"
version = "0.1.0"
description = ""
authors = ["Seu Nome <jackson.pacheco.taking@ingresse.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.37.1"
google-generativeai = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
